Added to GitHub
A stock statement generator:

actions: the timestamped actions that the stock trader performed, it can be BUY or SELL type, and they can buy or sell a few different stocks.

stock_actions: timestamped actions that the stock performed regardless of who the trader is. It includes stock splits, and dividend payouts.